requirements.txt
flask
python-dotenv
requests
beautifulsoup4
selenium
rsa
schedule